+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- A toggle button for copying non-unique patterns before editing. When enabled
+  and if the pattern is used in multiple places, the pattern is copied first.
+  ([#77][i77])
 - User can define own keybindings in
   `os.UserConfigDir()/sointu/keybindings.yaml` ([#94][i94], [#151][i151])
 - A small number above the instrument name identifies the MIDI channel /
@@ -226,6 +229,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [0.1.0]: https://github.com/vsariola/sointu/compare/4klang-3.11...v0.1.0
 [i65]: https://github.com/vsariola/sointu/issues/65
 [i68]: https://github.com/vsariola/sointu/issues/68
+[i77]: https://github.com/vsariola/sointu/issues/77
 [i94]: https://github.com/vsariola/sointu/issues/94
 [i112]: https://github.com/vsariola/sointu/issues/112
 [i116]: https://github.com/vsariola/sointu/issues/116